修改数据类型

+++ server/src/data/dto.rs
@@ -3,5 +3,5 @@ use serde::Serialize;
 #[derive(Serialize)]
 pub struct DataDto {
     pub time: String,
-    pub value: f64
+    pub value: i32
 }
